Two changes, 1 small and 1 big to optimise changePizzaSizes function computation
The small change is that I replaced querySelectorAll with
getElementsByClass. They both do the same but the latter might require
the browser to do more work as it needs to prepare for a wider range of
parameter value type, whereas the other is more specific, it only
accepts classes as parameter.
The big change is I noticed I can seperate the computation part which
only needs to be looped once as the values of dx and  newwidth always
stays the same no matter how many times we calculate it. Once we have
both values we can use the in the second loop to set the newwidth to all
the pizza elements. This drastically improved the performance without
changing the function result.