large optimisation of updatePosition function
This optimisation altered the code quite a bit. From reading in the
forums the variable phase has 5 unique values only. So instead I put the
phase variable into a different loop where it can store those unique
values into an Array. I than added this to the main loop event and
iterated it 5 times before I reset it to 0 and over again as beforehand
that was the basic pattern of the values within the main equation. This
brought down the speed to 0.5ms to 0.8ms for last 10 frames to generate.
A far cry from the over 20ms I started with.