changePizzaSizes no longer uses determineDx
I removed the need to use determineDx which was used to calculate the
difference in sizes of the pizza and and pixel vs percentage value.
Thanks to some class videos I found out that we can directly input the
percentage value of our pizza as we select the slider value.